// src/app/contenido/interfaces.ts
export interface LanguageContent {
    en: string;
    de: string;
    es: string;
}

export type Language = "es" | "en" | "de";

export type Dioses = "" | "Afrodita" | "Apolo" | "Ares" | "Artemisa" | "Atena" | "Demeter" | "Hefesto" | "Hera" | "Hermes" | "Hestia" | "Poseidon" | "Zeus";
// ...
export interface PrinterContent {
    title: LanguageContent;
    languageselection: LanguageContent;
    cerrarsesion: LanguageContent;
    subtitulo: LanguageContent;
    hora: LanguageContent;
    nombre: LanguageContent;
    dios: LanguageContent;
    idioma: LanguageContent;
    archivo: LanguageContent;
    opciones: LanguageContent;
    visualizar: LanguageContent;
    eliminar: LanguageContent;
    cancelar: LanguageContent;
    estasseguro: LanguageContent;
    vermas: LanguageContent;
}